Add prefix to velero BackupStorageLocation
In order to prevent velero validation errors when the backup storage
bucket contains migration registry storage as well, the velero
backups need to be configured to use a directory prefix within
the bucket.

This commit hard-codes that prefix to "velero". A more flexible
approach would add a field to BackupStorageConfig, but that would
probably only be needed if there was some need to have two
completely different sets of velero backups in one bucket. Without that,
this should be sufficient.