Fail migrations for velero PartiallyFailed output
As of Velero 1.0, backups and restores can end with a "PartiallyFailed"
state in addition to just a "Failed" state.

While it's possible that some PartiallyFailed backups or restored are still
usable, the cases I've seen so far have been essentially failed for
migration purposes, so this commit handles PartiallyFailed in the same
way as Failed, although the error message does distinguish the two cases.

fixes #154