Add pod affinity selector to stage pod definition (#235)
Stage pods need to run on the same node as the related app pod, if possible.
Some storage types (glusterblock, for example) will cause the stage
pods to fail to start if they're on a different node as the app
pod. This PR defines a soft affinity, so it will go on the same
node if it is able to schedule it there, but if not, it will schedule
it elsewhere, limiting the failure scenario to only failing if the
desired node is unavailable *and* it's one of the affected storage
types.