Replace PodAffinity with explicit NodeName selection for stage pods (#327)
At the moment we don't handle the case where the selected NodeName
doesn't exist in the cluster. If it doesn't exist, then the pod will
disappear, which might confuse the controller into thinking it hasn't been
created yet. However, since the pod name is taken from the (currently
running) application pods, it's not clear whether this is actually possible,
except as a race condition where a migration is running while a node name
change is being processed, or if it's possible for a running pod to
*not* be updated of a previous NodeName change.