Return after setting error for creds validation to avoid nil segfault
After not finding the secret, the validation code was setting a critical
condition for the missing secret, and then going on to the next line
attempting to validate the (non-existent) secret, causing the controller
to crash.

This commit adds the appropriate return statements after setting critical
validation conditions related to the secret so that we don't attempt further
validation on the secret.