Limit velero pod search to expected velero namespace.
If velero exists in other namespaces using the upstream image, the
validation findPod call may fail if the upstream velero is encountered
last in the list since the upstream image does not include /usr/bin/nc.
Instead we need to filter the velero pod search to only look in the migration
namespace since we know that velero will be from our fork, containing nc.