First phase work for direct image migration controller
This creates a controller for DirectImageMigration which migrates internal imagestreams
from a set of namespaces on the src cluster to a set of namespaces on the dest cluster.
There's also a secondary controller, created by the first, which manages the migration
of internal images in a single imagestream from the source to destination clusters.

There is not yet any integration with the migplan/migmigration workflow.