change the dns safe name to md5 hash and add annotation
Before this, PVC with longer names was truncated with the first 63 characters.
This introduced a possibility of collisions where two PVCs with names sharing
a common set of the first 63 characters.

This uses a md5 hash instead of truncating the characters to avoid collision still
getting unique strings smaller than 63 chars to use. It add annotation for the
PVC name so as to aid debugging.