Bug 2021668: Respect cluster_subdomain setting
There was a regression where the cluster_subdomain key was not being
used when creating dvm routes. This PR allows the cluster_subdomain to
be overridden when creating routes.

Additionally, this PR grabs the subdomain from the OpenShift's cluster
ingress config (supported since 4.3
https://docs.openshift.com/container-platform/4.3/networking/ingress-operator.html),
allowing crane-lib's endpoint functionality to handle scenarios where
the route name would be too long.

Add the config API to the manager's scheme as that is automatically used
in migration cluster client schemes.