Ensure label selector matches both filesystem and block labels (#1390)
The block server pod had a slightly different label
than the filesystem server pod, and this caused the
code that waits for the server to be running to not
match the pod. This then caused the clients to start
while the server was not yet running. This leads to
a race condition where the clients can attempt to
connect to the server before it is ready to accept
the connections. This then leads to tls errors and
the transfer failing.

Signed-off-by: Alexander Wels <awels@redhat.com>