Instead -new add -mig-xxxx to the end of the new pv names (#1393)
In order to support multiple migrations of the same PV
we need to add a unique suffix to the end of the PV name
when creating new PVs. The new suffix is '-mig-' + 4 character
random string of letters. When migrating a PV that already
has the suffix, we will remove the suffix and create a new
one.

In order to support legacy plans and pvs if the suffix is
-new we replace -new with -mig-xxxx for the new pv name.

relaxed the validation of pvs to only include pvs that
are not being skipped in the plan. This allows for the
original pv to still exist in the namespace.

Signed-off-by: Alexander Wels <awels@redhat.com>