Handle VM state changes during storage live migration (#1394)
It is possible that during the storage live migration
of a VM the state of that VM changes. This can happen
when the VM is started or stopped. When a running
VM is stopped, the live migration is cancelled and
an offline migration is started instead.

The reverse is also true, if a VM is started during
and offline migration, the offline migration is
cancelled and a live migration is started instead.

Since offline migrations use a single rsync server
with potentially multiple clients. If we stop a VM
but another offline migration is running, that one
is allowed to complete before starting a new offline
migration. The reverse is also true, if a VM is started
while it is part of an offline migration, the offline
migration is allowed to complete before starting a
live migration.

Signed-off-by: Alexander Wels <awels@redhat.com>