Support VM volume storage migration between different volume and access modes. This allows the user to live migration VMs from filesystem to block and vice versa. It also allows the user to change the access mode to RWX or RWO depending on their needs.
When creating the target volumes, use a datavolume
to create the appropriate PVCs, this allows the user
to select the storage profile to pick the correct
combination for the selected storage class.

Added ownertType of PVs, this helps the UI display
the selection table correctly. Based on the ownerType
it can change the appearance of the table.

Signed-off-by: Alexander Wels <awels@redhat.com>