Fix starting a migration sometimes throwing a PVC not found error. It was cleaning stale VirtualMachineInstanceMigration objects and re-using a function that looked up PVCs that might not have been created yet. This changes the check to not look at PVCs at all.
Fixed an issue where completed virt-launcher pods
where not being cleaned up, which would sometimes
cause an issue on rollback where it would not
complete because the completed virt-launcher pod
was holding onto the PVC we wanted to delete. This
would cause the PVC to stick around and the rollback
would not complete because it was waiting for the
PVC to be deleted.

Signed-off-by: Alexander Wels <awels@redhat.com>